Rewrite most of the interface after a hard look on how the whole pmu stuff _really_ works.
I previously assumed that every event in perf/pmu-events is routed to
the "cpu" PMU. This is not the case. There is quite a lot of code to go
around for selecting the correct PMU instance (from
/sys/bus/event_source/devices) for a given PMU name coming from
pmu-events.

For more information, see the code, I have tried to exensively comment
it.